eforms_project/eforms/requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==3.1.25
contextlib2==0.5.5
Django==1.10.7
django-cors-headers==2.0.2
django-unused-media==0.1.6
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
drfdocs==0.0.11
dropbox==7.3.1
google-api-python-client==1.6.2
httplib2==0.10.3
kombu==4.0.2
MySQL-python==1.2.5
oauth2client==4.1.0
olefile==0.44
packaging==16.8
pdfkit==0.6.1
Pillow==4.1.1
pkg-resources==0.0.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pydns==2.3.6
PyDrive==1.3.1
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
PyYAML==3.12
raven==6.0.0
reportlab==3.4.0
requests==2.14.2
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
validate-email==1.3
vine==1.1.3
wkhtmltopdf==0.2
